Разбил js файлы на модули
